Created the terminal, compiled
Added a while loop in the setup that reads for commands in the serial.
Made the randomizing and setup as functions.
Created a function terminal_commands(String) that takes a command and works it out.
Added the functions [START, ENABLE, DISABLE, SET_TIME, GET_SEED, RANDOMIZE. SET_NEEDY, SET_HARDCORE, SET_MILLIS_TIMER]
created a function that toggles a module.